+++ sites/partners/lib/hooks.ts
@@ -136,3 +136,18 @@ export function useSingleFlaggedApplication(afsId: string) {
     error,
   }
 }
+
+export function useAmiChartList() {
+  const { amiChartsService } = useContext(AuthContext)
+  const fetcher = () => amiChartsService.list()
+
+  const { data, error } = useSWR(`${process.env.backendApiBase}/amiCharts`, fetcher)
+  console.log("data = ", data)
+  console.log("error = ", error)
+
+  return {
+    data,
+    loading: !error && !data,
+    error,
+  }
+}